python-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.18

# Auth (LDAP/AD + JWT)
ldap3==2.9.1
PyJWT==2.10.1
